Fix handleError around ErrorEvents
Sometimes, discord.js hands an ErrorEvent to the `error` event, not an Error.
This is due to how it interfaces with one of its upstream libraries, ws.

For now, check if there's an `error` property on the given parameter, and if
so, use the value of that as the error instead.